pyproject.toml
[project]
name = "bangla_news_digest"
version = "1.0"
description = "A news scraper that scrapes from all Bangla news platforms and provides summary bullet points"
authors = [
    {name = "Muntasir Mohammad"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.35.0,<5.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "fake-useragent (>=2.2.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "docxtpl (>=0.20.1,<0.21.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "celery[auth,librabbitmq,redis] (>=5.5.3,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
hypothesis = "^6.138.15"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.13.0"
mypy = "^1.18.1"
bandit = "^1.8.6"


[tool.black]
line-length = 150
target-version = ["py313"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 150
target-version = "py313"
lint.ignore = ["E203", "F401"]
exclude = [".git", "__pycache__", "venv"]

[tool.mypy]
strict = true
disallow_untyped_defs = true
disable_error_code = ["type-arg"]

[[tool.mypy.overrides]]
module = ["docxtpl.*", "undetected_chromedriver.*", "sklearn.*"]
ignore_missing_imports = true
